+++ manticore/utils/emulate.py
@@ -1,4 +1,3 @@
-from __future__ import absolute_import
 import logging
 import inspect
