+++ manticore/utils/nointerrupt.py
@@ -1,4 +1,3 @@
-from __future__ import absolute_import
 import signal
 import logging
